fix: Update middleware configuration to include API routes for Clerk authentication
- Add /api/:path* to middleware matcher to ensure Clerk middleware applies to API routes
- Fix route patterns in createRouteMatcher to use proper path-to-regexp syntax
- Update middleware logic to handle API routes separately from i18n routing
- Resolves Clerk auth() detection issues in API endpoints

This fixes the error: 'Clerk: auth() was called but Clerk can't detect usage of clerkMiddleware()'
