+++ src/lib/i18n.ts
@@ -1,7 +1,7 @@
 export const locales = ['en', 'es'] as const;
 export type Locale = typeof locales[number];
 
-export const defaultLocale: Locale = 'en';
+export const defaultLocale: Locale = 'es';
 
 export const localeNames: Record<Locale, string> = {
   en: 'English',
