+++ src/locales/en/common.json
@@ -45,9 +45,7 @@
     "subtotal": "Subtotal",
     "tax": "Tax",
     "discount": "Discount",
-    "features": "Features",
-    "all": "All",
-    "note": "Note"
+    "features": "Features"
   },
   "status": {
     "active": "Active",
@@ -99,5 +97,43 @@
     "unauthorized": "Unauthorized access",
     "forbidden": "Access forbidden",
     "serverError": "Server error"
+  },
+  "home": {
+    "hero": {
+      "title": "Level Up Your {valorant} Game",
+      "subtitle": "Complete challenging missions based on your real Valorant gameplay. Track your progress, earn rewards, and compete with other players.",
+      "cta": "Start Your Journey"
+    },
+    "features": {
+      "realTime": {
+        "title": "Real-Time Missions",
+        "description": "Complete missions like \"Get 10 headshots\" or \"Win 5 Spike Rush games\" based on your actual Valorant match data."
+      },
+      "rewards": {
+        "title": "Earn Rewards",
+        "description": "Gain points and unlock achievements as you complete missions. Track your progress and celebrate your victories."
+      },
+      "compete": {
+        "title": "Compete & Compare",
+        "description": "See how you stack up against other players on the leaderboard. Challenge friends and climb the ranks."
+      }
+    },
+    "howItWorks": {
+      "title": "How It Works",
+      "steps": {
+        "connect": {
+          "title": "Connect Your Account",
+          "description": "Link your Valorant account by providing your Riot ID and tag."
+        },
+        "choose": {
+          "title": "Choose Missions",
+          "description": "Select from various missions tailored to your skill level and playstyle."
+        },
+        "play": {
+          "title": "Play & Progress",
+          "description": "Play Valorant normally and watch your mission progress update automatically."
+        }
+      }
+    }
   }
 }
