+++ src/locales/en/errors.json
@@ -1,4 +1,4 @@
-<<<<<<< HEAD
+
 {
   "api": {
     "networkError": "Network error. Please check your connection.",
