+++ src/locales/en/navigation.json
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 {
   "brand": {
     "name": "Valorant Missions",
