+++ src/locales/en/profile.json
@@ -1,4 +1,4 @@
-<<<<<<< HEAD
+
 {
   "title": "Profile Settings",
   "description": "Configure your profile and connect your Valorant account to start tracking missions.",
