+++ src/locales/en/setup.json
@@ -1,4 +1,4 @@
-<<<<<<< HEAD
+
 {
   "title": "Welcome to Valorant Missions!",
   "description": "To start tracking your missions and progress, we need to connect your Valorant account.",
