+++ src/locales/es/dashboard.json
@@ -1,4 +1,4 @@
-<<<<<<< HEAD
+
 {
   "loading": "Cargando...",
   "loadingDashboard": "Cargando panel...",
