+++ src/locales/es/errors.json
@@ -1,4 +1,4 @@
-<<<<<<< HEAD
+
 {
   "api": {
     "networkError": "Error de red. Por favor verifica tu conexión.",
