+++ src/locales/es/missions.json
@@ -1,40 +1,36 @@
+
 {
+  "types": {
+    "kills": "eliminaciones",
+    "headshots": "disparos a la cabeza",
+    "gamemode": "modo de juego",
+    "weapon": "arma",
+    "rounds": "rondas",
+    "wins": "victorias"
+  },
   "difficulty": {
     "easy": "Fácil",
     "medium": "Medio",
     "hard": "Difícil"
-  },
-  "types": {
-    "kills": "Eliminaciones",
-    "headshots": "Disparos a la cabeza",
-    "gamemode": "Modo de juego",
-    "weapon": "Arma",
-    "rounds": "Rondas",
-    "wins": "Victorias"
-  },
-  "status": {
-    "available": "Disponible",
-    "active": "Activa",
-    "completed": "Completada",
-    "expired": "Expirada"
   },
   "filters": {
     "searchPlaceholder": "Buscar misiones...",
     "filterBy": "Filtrar por",
     "difficulty": "Dificultad",
     "type": "Tipo",
-    "status": "Estado",
-    "clearFilters": "Limpiar filtros",
-    "showingResults": "Mostrando {{count}} de {{total}} misiones",
-    "noResults": "No se encontraron misiones que coincidan con tus filtros",
-    "upgradeForFilters": {
-      "free": "Mejora a Estándar o Premium para acceder a filtros de misiones y búsqueda.",
-      "standard": "Mejora a Premium para acceder a filtros avanzados de misiones y búsqueda."
-    }
+    "clearFilters": "Limpiar Filtros",
+    "showingResults": "Mostrando {count} de {total} misiones",
+    "noResults": "No se encontraron misiones que coincidan con tus criterios"
   },
   "actions": {
     "acceptMission": "Aceptar Misión",
     "viewProgress": "Ver Progreso",
     "completeMission": "Completar Misión"
+  },
+  "status": {
+    "available": "Disponible",
+    "active": "Activa",
+    "completed": "Completada",
+    "expired": "Expirada"
   }
 }
