+++ src/locales/es/navigation.json
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 {
   "brand": {
     "name": "Misiones Valorant",
