+++ src/locales/es/profile.json
@@ -1,4 +1,4 @@
-<<<<<<< HEAD
+
 {
   "title": "Configuración del Perfil",
   "description": "Configura tu perfil y conecta tu cuenta de Valorant para comenzar a rastrear misiones.",
