+++ src/locales/es/setup.json
@@ -1,4 +1,4 @@
-<<<<<<< HEAD
+
 {
   "title": "¡Bienvenido a Valorant Missions!",
   "description": "Para comenzar a rastrear tus misiones y progreso, necesitamos conectar tu cuenta de Valorant.",
